商品订单会员商城配置等模块mp改造

+++ yshop-shop/src/main/java/co/yixiang/modules/shop/service/dto/OrderTimeDataDTO.java
@@ -21,8 +21,8 @@ public class OrderTimeDataDTO implements Serializable {
     private Integer lastWeekCount;//上周
     private Double lastWeekPrice; //上周
 
-    private Long userCount;
-    private Long orderCount;
+    private Integer userCount;
+    private Integer orderCount;
     private Double priceCount;
-    private Long goodsCount;
+    private Integer goodsCount;
 }
